Close dashboard auth boundary follow-up
