+++ packages/cli/src/daemon/route-request-identity.ts
@@ -25,8 +25,9 @@ export function resolveRequestPrincipal(
 export function resolveRouteRequestIdentity(
   req: IncomingMessage,
   agent: RouteIdentityAgent,
+  explicitRequestAuth?: RequestAuthContext,
 ): RouteRequestIdentity {
-  const requestAuth = getRequestAuthContext(req);
+  const requestAuth = explicitRequestAuth ?? getRequestAuthContext(req);
   const credentialToken = requestAuth?.source === "dashboard-session"
     ? requestAuth.internalCredentialToken
     : requestAuth?.token ?? extractBearerToken(req.headers.authorization);
